+++ TelnyxWebRTCDemo/ViewControllers/HomeViewController.swift
@@ -21,6 +21,7 @@ class HomeViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+
         self.appDelegate.voipDelegate = self
         self.telnyxClient = self.appDelegate.telnyxClient
         
@@ -175,11 +176,13 @@ extension HomeViewController {
 extension HomeViewController: SipCredentialsViewControllerDelegate {
     func onNewSipCredential(credential: SipCredential?) {
         let deviceToken = userDefaults.getPushToken()
-        guard let sipCredential = credential else {
-            print("HomeVeiwController :: connectButtonTapped() ERROR: SIP User and Password should not be empty.")
-            return
+        if let newProfile = credential {
+            if newProfile.isToken ?? false {
+                connectToTelnyx(telnyxToken: newProfile.username, sipCredential: nil, deviceToken: deviceToken)
+            } else {
+                connectToTelnyx(telnyxToken: nil, sipCredential: newProfile, deviceToken: deviceToken)
+            }
         }
-        connectToTelnyx(telnyxToken: nil, sipCredential: sipCredential, deviceToken: deviceToken)
     }
     
     func onSipCredentialSelected(credential: SipCredential?) {
@@ -288,7 +291,9 @@ extension HomeViewController {
                                 // You can choose the appropriate verbosity level of the SDK.
                                 logLevel: .all,
                                 // Enable webrtc stats debug
-                                debug: true)
+                                debug: true,
+                                // Force relay candidate
+                                forceRelayCandidate: false)
         } else if let credential = sipCredential {
             // To obtain SIP credentials, please go to https://portal.telnyx.com
             txConfig = TxConfig(sipUser: credential.username,
@@ -300,11 +305,15 @@ extension HomeViewController {
                                 logLevel: .all,
                                 reconnectClient: true,
                                 // Enable webrtc stats debug
-                                debug: true)
+                                debug: true,
+                                // Force relay candidate.
+                                forceRelayCandidate: false)
             
             // Store user / password in user defaults
             SipCredentialsManager.shared.addOrUpdateCredential(credential)
             SipCredentialsManager.shared.saveSelectedCredential(credential)
+            // Update UI
+            self.onSipCredentialSelected(credential: credential)
         }
         
         guard let config = txConfig else {